# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: No-Backdoor System Architecture API
  description: |
    Security verification platform backend API. Provides endpoints for dashboard metrics,
    evidence upload, system portfolio management, and verification queue operations.
  version: 1.0.0
  contact:
    name: API Support
    email: api@nobackdoor.local
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.nobackdoor.local/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /evidence/{id}:
    get:
      tags: [Evidence Upload]
      summary: Get upload detail
      description: Retrieve details of a single evidence upload.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "evd_abc123"
      responses:
        '200':
          description: Upload detail retrieved
          content:
            application/json:
              schema:
                type: object
                required: [success, data, meta]
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/EvidenceUpload'
                  meta:
                    $ref: '#/components/schemas/Meta'
              example:
                success: true
                data:
                  id: "evd_abc123"
                  fileName: "source-review.pdf"
                  originalName: "source-review.pdf"
                  mimeType: "application/pdf"
                  size: 2457600
                  status: "completed"
                  systemId: "sys_fw3"
                  systemName: "Firewall-v3"
                  description: "Source code review documentation for v2.4.1"
                  tags: ["source-code", "v2.4.1", "review"]
                  uploadedBy: "usr_abc123"
                  uploaderName: "Jane Doe"
                  downloadUrl: "https://cdn.example.com/evidence/evd_abc123.pdf"
                  createdAt: "2024-01-15T14:20:00Z"
                  updatedAt: "2024-01-15T14:20:00Z"
                meta:
                  timestamp: "2024-01-15T14:20:00Z"
                  requestId: "req_cde234"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Evidence Upload]
      summary: Delete an upload
      description: Remove an evidence upload and its associated file.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "evd_abc123"
      responses:
        '204':
          description: Upload deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================================
  # Portfolio (Systems)
  # ============================================================
  /systems:
    get:
      tags: [Portfolio]
      summary: List all systems
      description: |
        Retrieve a paginated, searchable, and filterable list of systems.
        Supports filtering by status, type, and free-text search.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          schema:
            type: string
            minLength: 1
            maxLength: 200
          description: Search by name, description, or owner
        - name: status
          in: query
          schema:
            type: string
            enum: [verified, pending, failed, not_started, in_progress]
          description: Filter by verification status
        - name: type
          in: query
          schema:
            type: string
            enum: [web_app, mobile_app, api, database, network, desktop, cloud_service, embedded, other]
          description: Filter by system type
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, createdAt, updatedAt, status, lastVerifiedAt]
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: ownerId
          in: query
          schema:
            type: string
          description: Filter by system owner
      responses:
        '200':
          description: Systems list retrieved
          content:
            application/json:
              schema:
                type: object
                required: [success, data, meta]
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/System'
                  meta:
                    allOf:
                      - $ref: '#/components/schemas/Meta'
                      - $ref: '#/components/schemas/PaginationMeta'
              example:
                success: true
                data:
                  - id: "sys_fw3"
                    name: "Firewall-v3"
                    description: "Next-gen firewall with ML-based threat detection"
                    type: "network"
                    status: "verified"
                    ownerId: "usr_abc123"
                    ownerName: "Jane Doe"
                    version: "3.2.1"
                    lastVerifiedAt: "2024-01-15T10:00:00Z"
                    evidenceCount: 12
                    tags: ["security", "network", "ml"]
                    createdAt: "2023-06-01T00:00:00Z"
                    updatedAt: "2024-01-15T10:00:00Z"
                  - id: "sys_pg1"
                    name: "Payment Gateway"
                    description: "PCI-DSS compliant payment processing service"
                    type: "api"
                    status: "pending"
                    ownerId: "usr_def456"
                    ownerName: "John Smith"
                    version: "2.0.0"
                    lastVerifiedAt: null
                    evidenceCount: 5
                    tags: ["financial", "pci-dss", "api"]
                    createdAt: "2023-08-15T00:00:00Z"
                    updatedAt: "2024-01-10T16:30:00Z"
                meta:
                  timestamp: "2024-01-15T14:20:00Z"
                  requestId: "req_def345"
                  pagination:
                    page: 1
                    limit: 20
                    total: 142
                    totalPages: 8
                    hasNext: true
                    hasPrev: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Portfolio]
      summary: Create new system
      description: Add a new system to the portfolio catalog.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSystemRequest'
            example:
              name: "Auth Service v2"
              description: "OAuth 2.0 and SAML authentication microservice"
              type: "api"
              ownerId: "usr_abc123"
              version: "2.1.0"
              tags: ["auth", "oauth", "saml"]
              metadata:
                repository: "https://git.example.com/auth-service"
                documentation: "https://docs.example.com/auth"
      responses:
        '201':
          description: System created successfully
          content:
            application/json:
              schema:
                type: object
                required: [success, data, meta]
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/System'
                  meta:
                    $ref: '#/components/schemas/Meta'
              example:
                success: true
                data:
                  id: "sys_as2"
                  name: "Auth Service v2"
                  description: "OAuth 2.0 and SAML authentication microservice"
                  type: "api"
                  status: "not_started"
                  ownerId: "usr_abc123"
                  ownerName: "Jane Doe"
                  version: "2.1.0"
                  lastVerifiedAt: null
                  evidenceCount: 0
                  tags: ["auth", "oauth", "saml"]
                  metadata:
                    repository: "https://git.example.com/auth-service"
                    documentation: "https://docs.example.com/auth"
                  createdAt: "2024-01-15T14:25:00Z"
                  updatedAt: "2024-01-15T14:25:00Z"
                meta:
                  timestamp: "2024-01-15T14:25:00Z"
                  requestId: "req_efg456"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: System name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                success: false
                error:
                  code: "CONFLICT"
                  message: "System name already exists"
                  status: 409
                meta:
                  timestamp: "2024-01-15T14:25:00Z"
                  requestId: "req_efg456"
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/refresh
  schemas:
...
    Meta:
      type: object
      required: [timestamp, requestId]
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    PaginationMeta:
      type: object
      required: [pagination]
      properties:
        pagination:
          type: object
          required: [page, limit, total, totalPages, hasNext, hasPrev]
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
            hasNext:
              type: boolean
            hasPrev:
              type: boolean
    ApiError:
      type: object
      required: [success, error, meta]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message, status]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid request parameters"
            details:
              type: string
              description: Additional error details
              example: "Field 'email' must be a valid email address"
            status:
              type: integer
              description: HTTP status code
              example: 400
        meta:
          $ref: '#/components/schemas/Meta'

    # ----------------------------------------------------------
    # Auth & Users
    # ----------------------------------------------------------
...
    UploadStatus:
      type: string
      enum: [pending, processing, completed, failed]
    EvidenceUpload:
      type: object
      required: [id, fileName, originalName, mimeType, size, status, uploadedBy, createdAt, updatedAt]
      properties:
        id:
          type: string
        fileName:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
          description: File size in bytes
        status:
          $ref: '#/components/schemas/UploadStatus'
        systemId:
          type: string
          nullable: true
        systemName:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
          nullable: true
        uploadedBy:
          type: string
        uploaderName:
          type: string
        downloadUrl:
          type: string
          format: uri
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ----------------------------------------------------------
    # Portfolio (Systems)
    # ----------------------------------------------------------
    SystemType:
      type: string
      enum: [web_app, mobile_app, api, database, network, desktop, cloud_service, embedded, other]
    SystemStatus:
      type: string
      enum: [verified, pending, failed, not_started, in_progress]
    System:
      type: object
      required: [id, name, description, type, status, ownerId, ownerName, version, lastVerifiedAt, evidenceCount, tags, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          $ref: '#/components/schemas/SystemType'
        status:
          $ref: '#/components/schemas/SystemStatus'
        ownerId:
          type: string
        ownerName:
          type: string
        version:
          type: string
        lastVerifiedAt:
          type: string
          format: date-time
          nullable: true
        evidenceCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSystemRequest:
      type: object
      required: [name, description, type]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          minLength: 1
          maxLength: 2000
        type:
          $ref: '#/components/schemas/SystemType'
        ownerId:
          type: string
          description: Defaults to current user if not provided
        version:
          type: string
          maxLength: 50
          default: "1.0.0"
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 20
        metadata:
          type: object
          additionalProperties: true
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "Invalid request parameters"
              details: "Field 'email' must be a valid email address"
              status: 400
            meta:
              timestamp: "2024-01-15T14:20:00Z"
              requestId: "req_err001"
    Unauthorized:
      description: Unauthorized - missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
              details: "Bearer token is missing or invalid"
              status: 401
            meta:
              timestamp: "2024-01-15T14:20:00Z"
              requestId: "req_err002"
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: "FORBIDDEN"
              message: "Insufficient permissions"
              details: "Admin role required for this operation"
              status: 403
            meta:
              timestamp: "2024-01-15T14:20:00Z"
              requestId: "req_err003"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
              details: "The requested resource does not exist"
              status: 404
            meta:
              timestamp: "2024-01-15T14:20:00Z"
              requestId: "req_err004"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: "INTERNAL_ERROR"
              message: "An unexpected error occurred"
              status: 500
            meta:
              timestamp: "2024-01-15T14:20:00Z"
              requestId: "req_err005"